requirements.txt
Django==3.2.3
django-environ==0.4.5
djangorestframework==3.12.4
drf-spectacular==0.18.1
psycopg2-binary==2.9.1
Pillow==8.3.1
djangorestframework-simplejwt==4.8.0
PyJWT==2.1.0
django-storages==1.11.1
boto3==1.18.40
botocore==1.21.40
boto==2.49.0
pytest==6.2.5
pytest-faker==2.0.0
pytest-factoryboy==2.1.0
pytest-sugar==0.9.4
factories==1.2.0
factory-boy==3.2.0
Faker==8.9.1
requests==2.26.0
